requirements.txt
# Core Framework
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.0.2

# PDF Processing
PyPDF2>=3.0.1
pdfplumber>=0.9.0
pymupdf>=1.23.0

# Vector Databases
faiss-cpu>=1.7.4
chromadb>=0.4.15

# OpenAI
openai>=1.3.0

# Text Processing
tiktoken>=0.5.1
sentence-transformers>=2.2.2

# Data Handling
pandas>=2.0.0
numpy>=1.24.0

# Environment Management
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Async support
aiofiles>=23.2.0
asyncio>=3.4.3 